eight  commit

+++ src/main/java/com/driver/services/impl/ParkingLotServiceImpl.java
@@ -55,34 +55,39 @@ public class ParkingLotServiceImpl implements ParkingLotService {
 
     @Override
     public Spot updateSpot(int parkingLotId, int spotId, int pricePerHour) {
-        Spot spot = spotRepository1.findById(spotId).get();
-        spot.setPricePerHour(pricePerHour);
-        ParkingLot parkingLot = parkingLotRepository1.findById(parkingLotId).get();
-        List<Spot> spotList = parkingLot.getSpotList();
-        if(spotList == null){
-            spotList = new ArrayList<>();
-            spotList.add(spot);
-        }
-        for (Spot spots: spotList ) {
-            if(spots.getId()== spotId){
-                spots.setPricePerHour(pricePerHour);
+        Spot spot = null;
+        if(spotRepository1.findById(spotId).get() != null) {
+           spot = spotRepository1.findById(spotId).get();
+            spot.setPricePerHour(pricePerHour);
+            ParkingLot parkingLot = parkingLotRepository1.findById(parkingLotId).get();
+            List<Spot> spotList = parkingLot.getSpotList();
+            if (spotList == null) {
+                spotList = new ArrayList<>();
+                spotList.add(spot);
             }
+            for (Spot spots : spotList) {
+                if (spots.getId() == spotId) {
+                    spots.setPricePerHour(pricePerHour);
+                }
+            }
+            spot.setParkingLot(parkingLot);
+            parkingLotRepository1.save(parkingLot);
+            spotRepository1.save(spot);
+
         }
-        spot.setParkingLot(parkingLot);
-        parkingLotRepository1.save(parkingLot);
-        spotRepository1.save(spot);
         return spot;
     }
 
     @Override
     public void deleteParkingLot(int parkingLotId) {
-        ParkingLot parkingLot = parkingLotRepository1.findById(parkingLotId).get();
 
-            List<Spot> spotList = parkingLot.getSpotList();
-            for (Spot spot : spotList) {
-                deleteSpot(spot.getId());
+            if(parkingLotRepository1.findById(parkingLotId).get()!= null) {
+                ParkingLot parkingLot = parkingLotRepository1.findById(parkingLotId).get();
+                List<Spot> spotList = parkingLot.getSpotList();
+                for (Spot spot : spotList) {
+                    deleteSpot(spot.getId());
+                }
+                parkingLotRepository1.deleteById(parkingLotId);
             }
-            parkingLotRepository1.deleteById(parkingLotId);
-
     }
 }
